memory optimizations

+++ wan/modules/vae.py
@@ -621,7 +621,7 @@ class WanVAE:
     def __init__(self,
                  z_dim=16,
                  vae_pth='cache/vae_step_411000.pth',
-                 dtype=torch.float,
+                 dtype=torch.bfloat16,
                  device="cuda"):
         self.dtype = dtype
         self.device = device
